Mainly adds 2 tests (tehy should not pass yet though)
One test is for the clock (clock-test.cc). It sets a frequency and checks that the
clock time matches the tested skew against simulator time.
It is possible to change the frequency during the test to check that it
calls the test callback. It should also be possible to test against
timesteps but this is not implemented yet.

The 2nd test is in "node-schedule-test" and is equivalent to simulator-test-suite but with calls like m_node->Schedule( Time(3), ...).
It also checks that events trigger according to node's clock.
Node does not the Destroy functions because it doesn't seem fundamental.

As a way to quickly test the API, I've implemented some members with
variadic templates thus you would need:
CXXFLAGS="  -std=c++11" ./waf configure