Decorrelate SND.UNA from Tx buffer
ns3 considers that as soon as in order data is acked, it can be removed
from the buffer. In this situation, the TcpTxBuffer::HeadSequence is
equivalent to SND.UNA.
For MpTcp, things are different since before being able to free date
from the Tx buffer, it needs to be acknowledged at both the TCP and
MPTCP levels.