Various test.py improvements
Upgrading to python3 argparse.
You can now chain several testsuites:
/test.py -stcp-rto-test -stcp-option

You can now pass additionnal arguments such as
./test.py -etcp-large-transfer --PrintGlobals

Also constrains some parameters with argparse in order to fail early,
without calling test-runner.

Addition of a new flag -D to display debug info that for now does the
same as --verbose with the hope that in the future --verbose can be
passed to tests as in the test-runner, in order to display more detailed
diagnostics.

The order of ./test.py -l has changed.